+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.8 Locale
+- Add original message to MessageContext
+- When receiving a command pass text via MessageContext.text field
+
 ## 1.0.7 Locale
 - Add user locale to MessageContext